language_prediction/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
albumentations==0.5.2
astral==2.2
astropy==5.1
attrdict==2.0.1
backports.strenum
beartype==0.11.0
CairoSVG==2.5.2
colorlog==6.6.0
coverage==6.2
cstruct==3.1
cycler==0.10.0
dask==2022.9.1
decorator==4.4.2
deepdiff==5.8.0
einops==0.4.1
geopandas==0.12.2
gitpython==2.0.0
hyperopt==0.2.7
ijson==3.1.4
imageio==2.9.0
imgaug==0.4.0
jinja2==3.0.3
jira==3.4.1
joblib==1.1.1
jsonpath-ng==1.5.3
kiwisolver==1.3.1
kornia==0.6.5
laspy==2.0.3
laspy[laszip]
laszip==0.1.0
matplotlib==3.5.1
more-itertools==8.14.0
navitrainingtools==5.0.3
networkx==3.0
nptyping==2.3.1
numba==0.56.4
numpy==1.23.5
numpy-quaternion==2022.4.3
omegaconf==2.2.3
onnx==1.13.1
onnx-simplifier==0.3.6
open3d==0.16.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
openpyxl==3.1.1
orjson==3.8.1
pandas==1.4.3
pathos==0.2.9
pickle5
Pillow==9.5.0
protobuf==3.20.3
psutil==5.9.0
pyparsing==2.4.7
pyproj==3.4.1
pytest==7.0.1
pytest-cov==3.0.0
pytest-mock==3.7.0
pytest-timeout==2.1.0
python-dateutil==2.8.1
python-gitlab==3.13.0
python-jenkins==1.8.0
python-vlc==3.0.16120
pytorch-metric-learning
PyTurboJPEG
PyWavelets==1.4.1
PyYAML==6.0
rdp==0.8
requests==2.27.1
Rtree>=0.9.7
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.11.2
sensor-models>=0.74
separator_taxonomy~=0.2
Shapely==2.0.1
six==1.16.0
tensorboard==2.13.*
tensordict
threadpoolctl==3.0.0
tifffile==2020.9.3
torch==2.0.1+cu118
torchaudio==2.0.2
torchgeometry==0.1.2
torchmetrics==0.11.*
torchvision==0.15.2+cu118
tqdm==4.62.3
typeguard==2.13.3
visdom==0.1.8.9
datasets
tokenizers
altair
hydra-core
